feat: Implemented base bot functions

+++ app/pkg/database/models/threads.go
@@ -2,7 +2,7 @@ package models
 
 type Thread struct {
 	ID int `pg:",pk,notnull,autoincrement"`
-	ThreadID int64 `pg:",notnull"`
+	ThreadID int `pg:",notnull"`
 	CreatedAt int64 `pg:"default:extract(epoch from now())"`
 	UpdatedAt int64 `pg:"default:extract(epoch from now())"`
